+++ ribs/emitters/_evolution_strategy_emitter.py
@@ -103,7 +103,7 @@ class EvolutionStrategyEmitter(EmitterBase):
                          np.random.SeedSequence(seed))
         opt_seed, ranker_seed = seed_sequence.spawn(2)
 
-        self._x0 = np.array(x0, dtype=archive.dtype)
+        self._x0 = np.array(x0, dtype=archive.dtypes["solution"])
         check_shape(self._x0, "x0", archive.solution_dim,
                     "archive.solution_dim")
         self._sigma0 = sigma0
@@ -125,7 +125,7 @@ class EvolutionStrategyEmitter(EmitterBase):
             batch_size=batch_size,
             solution_dim=self._solution_dim,
             seed=opt_seed,
-            dtype=self.archive.dtype,
+            dtype=self.archive.dtypes["solution"],
             lower_bounds=self.lower_bounds,
             upper_bounds=self.upper_bounds,
             **(es_kwargs if es_kwargs is not None else {}),